+++ lib/rodauth/rails/feature/internal_request.rb
@@ -5,7 +5,7 @@ module Rodauth
         def domain
           return super unless missing_host?
 
-          rails_url_options[:host]
+          rails_url_options.fetch(:host)
         end
 
         def base_url
@@ -43,7 +43,8 @@ module Rodauth
         end
 
         def rails_url_options
-          ::Rails.application.config.action_mailer.default_url_options || {}
+          ::Rails.application.config.action_mailer.default_url_options or
+            fail Error, "There is no information to set the URL host from. Please set config.action_mailer.default_url_options in your Rails application, or configure #domain and #base_url in your Rodauth configuration."
         end
       end
     end